+++ src/factory.ts
@@ -1,9 +1,13 @@
-type AttributeBuilderParams<Entity, TransientParams> = {
-  sequence: number;
-  params: Partial<Entity>;
-  transientParams?: TransientParams;
+type EntityAttributes<Entity, TransientParams> = {
+  [Property in keyof Entity]: EntityAttribute<
+    Entity,
+    TransientParams,
+    Property
+  >;
 };
-
+type EntityAttribute<Entity, TransientParams, Property extends keyof Entity> =
+  | AttributeBuilder<Entity, TransientParams, Property>
+  | Entity[Property];
 type AttributeBuilder<
   Entity,
   TransientParams,
@@ -11,72 +15,283 @@ type AttributeBuilder<
 > = (
   params: AttributeBuilderParams<Entity, TransientParams>
 ) => Entity[Property];
-
-type SchemaAttribute<Entity, TransientParams, Property extends keyof Entity> =
-  | AttributeBuilder<Entity, TransientParams, Property>
-  | Entity[Property];
-
-type Schema<Entity, TransientParams> = {
-  [Property in keyof Entity]: SchemaAttribute<
-    Entity,
-    TransientParams,
-    Property
-  >;
+type AttributeBuilderParams<Entity, TransientParams> = {
+  sequence: number;
+  params: Partial<Entity>;
+  transientParams: TransientParams;
 };
 
-interface FactoryOptions<Entity, TransientParams> {
-  afterCreate?(
-    entity: Entity,
-    params: { transientParams?: TransientParams }
-  ): void;
+type TransientParamsForTraits<
+  Traits,
+  TraitNames extends keyof Traits
+> = UnionToIntersection<Pick<Traits, TraitNames>[TraitNames]>;
+type UnionToIntersection<U> = {
+  [K in GetKeys<U>]: U extends Record<K, infer T> ? T : never;
+};
+type GetKeys<U> = U extends Record<infer K, any> ? K : never;
+
+type FinalBuilder<Builder> = Omit<Builder, 'attributes' | 'transient'>;
+
+interface TraitBuilder<
+  Entity,
+  GlobalTransientParams = unknown,
+  TransientParams = {} // eslint-disable-line @typescript-eslint/ban-types
+> {
+  transient<TraitTransientParams extends Record<string, unknown>>(
+    params: TraitTransientParams
+  ): Omit<
+    TraitBuilder<Entity, GlobalTransientParams, TraitTransientParams>,
+    'transient'
+  >;
+
+  attributes(
+    attributes: EntityAttributes<
+      Partial<Entity>,
+      GlobalTransientParams & TransientParams
+    >
+  ): FinalBuilder<TraitBuilder<Entity, GlobalTransientParams, TransientParams>>;
+
+  afterCreate(
+    afterCreateCallback: (
+      entity: Entity,
+      params: { transientParams: GlobalTransientParams & TransientParams }
+    ) => void
+  ): FinalBuilder<TraitBuilder<Entity, GlobalTransientParams, TransientParams>>;
 }
 
-export class Factory<Entity, TransientParams = Record<string, any>> {
-  private schema: Schema<Entity, Partial<TransientParams>>;
-  private sequence = 0;
-  private options: FactoryOptions<Entity, Partial<TransientParams>>;
+interface FactoryBuilder<
+  Entity = unknown,
+  GlobalTransientParams = unknown,
+  Traits = unknown
+> {
+  transient<GlobalTransientParams>(
+    params: GlobalTransientParams
+  ): Omit<FactoryBuilder<Entity, GlobalTransientParams, Traits>, 'transient'>;
 
-  constructor(
-    schema: Schema<Entity, Partial<TransientParams>>,
-    options: FactoryOptions<Entity, Partial<TransientParams>> = {}
-  ) {
-    this.schema = schema;
-    this.options = options;
-  }
+  attributes<Entity>(
+    attributes: EntityAttributes<Entity, GlobalTransientParams>
+  ): FinalBuilder<FactoryBuilder<Entity, GlobalTransientParams, Traits>>;
 
-  build(params?: Partial<Entity>, transientParams?: Partial<TransientParams>) {
-    const sequence = ++this.sequence;
-    const result: Partial<Entity> = {};
-    for (let key in this.schema) {
-      let attribute = this.schema[key];
-      if (typeof attribute === 'function') {
-        result[key] = attribute({
-          sequence,
-          params: { ...result, ...params },
-          transientParams,
-        });
-      } else {
-        result[key] = attribute as any;
-      }
-    }
-    const entity = { ...result, ...params } as Entity;
-    this.options?.afterCreate?.(entity, { transientParams });
-    return entity;
-  }
+  trait<TraitName extends string, TraitTransientParams>(
+    name: TraitName,
+    traitBuilderCallBack: (
+      builder: TraitBuilder<Entity, GlobalTransientParams>
+    ) => FinalBuilder<
+      TraitBuilder<Entity, GlobalTransientParams, TraitTransientParams>
+    >
+  ): FinalBuilder<
+    FactoryBuilder<
+      Entity,
+      GlobalTransientParams,
+      Traits & Record<TraitName, TraitTransientParams>
+    >
+  >;
+
+  afterCreate(
+    afterCreateCallback: (
+      entity: Entity,
+      params: { transientParams: GlobalTransientParams }
+    ) => void
+  ): FinalBuilder<FactoryBuilder<Entity, GlobalTransientParams, Traits>>;
+}
+
+export interface EntityFactory<Entity, GlobalTransientParams, Traits> {
+  build(entity?: Partial<Entity & GlobalTransientParams>): Entity;
+
+  build<TraitNames extends keyof Traits>(
+    ...params:
+      | [
+          ...traits: [TraitNames, ...Array<TraitNames>],
+          params: Partial<
+            Entity &
+              GlobalTransientParams &
+              TransientParamsForTraits<Traits, TraitNames>
+          >
+        ]
+      | Array<keyof Traits>
+  ): Entity;
 
   buildList(
     count: number,
-    params?: Partial<Entity>,
-    transientParams?: Partial<TransientParams>
-  ) {
-    let result = [];
-    for (let i = 0; i < count; i++) {
-      result.push(this.build(params, transientParams));
-    }
-    return result;
-  }
+    entity?: Partial<Entity & GlobalTransientParams>
+  ): Array<Entity>;
 
-  rewindSequence() {
-    this.sequence = 0;
-  }
+  buildList<TraitNames extends keyof Traits>(
+    count: number,
+    ...params:
+      | [
+          ...traits: [TraitNames, ...Array<TraitNames>],
+          params: Partial<
+            Entity &
+              GlobalTransientParams &
+              TransientParamsForTraits<Traits, TraitNames>
+          >
+        ]
+      | Array<keyof Traits>
+  ): Array<Entity>;
+
+  rewindSequence(): void;
 }
+
+interface FactoryDefinition {
+  attributeDefaults: Record<string, any>;
+  transientParamDefaults: Record<string, any>;
+  traits: Record<
+    string,
+    {
+      attributeDefaults: Record<string, any>;
+      transientParamDefaults: Record<string, any>;
+      afterCreate?(...args: Array<any>): void;
+    }
+  >;
+  afterCreate(...args: Array<any>): void;
+  sequence: number;
+}
+
+export const Factory = {
+  define<Entity, GlobalTransientParams, Traits>(
+    factoryBuilderCallback: (
+      builder: FactoryBuilder
+    ) => FinalBuilder<FactoryBuilder<Entity, GlobalTransientParams, Traits>>
+  ): EntityFactory<Entity, GlobalTransientParams, Traits> {
+    const definition: FactoryDefinition = {
+      attributeDefaults: {},
+      transientParamDefaults: {},
+      traits: {},
+      afterCreate: () => null,
+      sequence: -1,
+    };
+
+    const factory: EntityFactory<Entity, GlobalTransientParams, Traits> = {
+      build(...args: any[]) {
+        let [params, ...traitNames] = args.reverse();
+        if (typeof params === 'string') {
+          traitNames.unshift(params);
+          params = {};
+        }
+
+        // split params into attributes and transient params
+        let attributes: Record<string, any> = {};
+        let transientParams: Record<string, any> = {
+          ...definition.transientParamDefaults,
+        };
+        for (let key in params) {
+          if (key in definition.attributeDefaults) {
+            attributes[key] = params[key];
+          } else {
+            transientParams[key] = params[key];
+          }
+        }
+
+        // build entity
+        const entity: Record<string, any> = attributes;
+        definition.sequence++;
+
+        // apply defaults from traits
+        for (let traitName of traitNames) {
+          const trait = definition.traits[traitName];
+
+          for (let key in trait.attributeDefaults) {
+            if (key in entity) continue;
+            let attribute = trait.attributeDefaults[key];
+            if (typeof attribute === 'function') {
+              entity[key] = attribute({
+                sequence: definition.sequence,
+                params: { ...attributes, ...entity },
+                transientParams: {
+                  ...trait.transientParamDefaults,
+                  ...transientParams,
+                },
+              });
+            } else {
+              entity[key] = attribute;
+            }
+          }
+        }
+
+        // apply global defaults
+        for (let key in definition.attributeDefaults) {
+          if (key in entity) continue;
+          let attribute = definition.attributeDefaults[key];
+          if (typeof attribute === 'function') {
+            entity[key] = attribute({
+              sequence: definition.sequence,
+              params: { ...attributes, ...entity },
+              transientParams,
+            });
+          } else {
+            entity[key] = attribute;
+          }
+        }
+
+        // after create hooks
+        for (let traitName of traitNames.reverse()) {
+          const trait = definition.traits[traitName];
+          trait.afterCreate?.(entity, {
+            transientParams: {
+              ...trait.transientParamDefaults,
+              ...transientParams,
+            },
+          });
+        }
+        definition.afterCreate(entity, { transientParams });
+
+        return entity as Entity;
+      },
+
+      buildList(count: number, ...args: any[]) {
+        let result = [];
+        for (let i = 0; i < count; i++) {
+          result.push(factory.build(...args));
+        }
+        return result;
+      },
+
+      rewindSequence() {
+        definition.sequence = -1;
+      },
+    };
+
+    const factoryBuilder: FactoryBuilder<any, any, any> = {
+      attributes(params) {
+        definition.attributeDefaults = params;
+        return factoryBuilder;
+      },
+      transient(params) {
+        definition.transientParamDefaults = params;
+        return factoryBuilder;
+      },
+      trait(name, traitBuilderCallBack) {
+        definition.traits[name] = {
+          attributeDefaults: {},
+          transientParamDefaults: {},
+          afterCreate: undefined,
+        };
+        const traitBuilder: TraitBuilder<any, any, any> = {
+          attributes(attributes) {
+            definition.traits[name].attributeDefaults = attributes;
+            return traitBuilder;
+          },
+          transient(params) {
+            definition.traits[name].transientParamDefaults = params;
+            return traitBuilder;
+          },
+          afterCreate(afterCreateCallback) {
+            definition.traits[name].afterCreate = afterCreateCallback;
+            return traitBuilder;
+          },
+        };
+        traitBuilderCallBack(traitBuilder);
+        return factoryBuilder;
+      },
+      afterCreate(afterCreateCallback) {
+        definition.afterCreate = afterCreateCallback;
+        return factoryBuilder;
+      },
+    };
+
+    factoryBuilderCallback(factoryBuilder);
+
+    return factory;
+  },
+};
